refactor(theme): remove ThemeToggle component and fix ThemeContext rendering
- Deleted ThemeToggle.tsx component and updated import path in Navigation.tsx
- Removed hydration mismatch prevention logic in ThemeProvider component
- Simplified ThemeProvider return to always render children within context provider
